python/requirements.txt
aiohttp==3.7.4.post0
colorama==0.4.4
discord.py==1.7.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
schedule==1.1.0
